backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
ffmpeg-python==0.2.0
websockets==12.0
python-dotenv==1.0.0
